Declare language in mailer layout, mark file name
My first email was sent to spam because google thought
it was in Portuguese. After testing what it wanted to 
translate, it was just the filename. In fact, all it
did was convert the date format (though it may also
have been space encoding) Either way, I should declare the
language and I’ll ask it not to translate that filename

https://webmasters.stackexchange.com/questions/106741/why-would-my-website-language-be-incorrectly-detected-and-how-do-i-set-it-to-th